Add 1 AI feature: Prompt to Form
I added one feature: Prompt to Form by LLM.

The way it works is actually simple. I added a backend (ai.py) that can handle query from user.

The query is sent to LLM (Kobold) for decomposition the case, LLM will send it back in JSON Format, and that JSON will be used as a reference by frontend to fill out the form.